refactor(recruit): reorganize lexical plugins, utils, editor into directories
