+++ sites/recruit.sveltevietnam.dev/src/lib/components/rich-text-editor/editor/plugins/floating-link-editor/index.ts
@@ -1,0 +1,2 @@
+export { default as FloatingLinkEditor } from './FloatingLinkEditor.svelte';
+export * from './register';
