+++ sites/recruit.sveltevietnam.dev/src/lib/components/rich-text-editor/editor/utils/get-selected-node.ts
