+++ sites/recruit.sveltevietnam.dev/src/lib/components/rich-text-editor/editor/utils/is-selecting-link.ts
